pyproject.toml
[project]
name = "foundation-anomaly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "transformers>=4.57.1",
]
